refactor: split monolithic controller into dedicated controllers
Replace the single OpenClawInstanceReconciler with two focused
controllers:
- OpenClawConfigMapReconciler: manages ConfigMap lifecycle
- OpenClawDeploymentReconciler: manages Deployment lifecycle (waits for
ConfigMap)

This separation of concerns improves maintainability and testability by
giving
each controller a single, clear responsibility. Both controllers watch the
same
OpenClawInstance resource but manage different owned resources.

Changes:
- Add OpenClawConfigMapController with ConfigMap lifecycle management
- Add OpenClawDeploymentController with Deployment lifecycle management
- Replace monolithic openclawinstance_controller.go with two focused
controllers
- Split test suite into dedicated test files per controller
- Create shared suite_test.go for common test infrastructure
- Update cmd/main.go to register both controllers
- Scope RBAC permissions per controller (ConfigMap controller: configmaps
r/w,
Deployment controller: configmaps read-only, deployments r/w)
- Archive completed split-controllers change

Architecture:
- Controllers operate independently (no inter-controller communication)
- Deployment controller checks for ConfigMap existence before creating
Deployment
- Maintains existing behavior: ConfigMap created first, then Deployment
- Each controller has focused RBAC permissions (principle of least
privilege)
