+++ cmd/main.go
@@ -202,11 +202,19 @@ func main() {
 		os.Exit(1)
 	}
 
-	if err = (&controller.OpenClawInstanceReconciler{
+	if err = (&controller.OpenClawConfigMapReconciler{
 		Client: mgr.GetClient(),
 		Scheme: mgr.GetScheme(),
 	}).SetupWithManager(mgr); err != nil {
-		setupLog.Error(err, "unable to create controller", "controller", "OpenClawInstance")
+		setupLog.Error(err, "unable to create controller", "controller", "OpenClawConfigMap")
+		os.Exit(1)
+	}
+
+	if err = (&controller.OpenClawDeploymentReconciler{
+		Client: mgr.GetClient(),
+		Scheme: mgr.GetScheme(),
+	}).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "OpenClawDeployment")
 		os.Exit(1)
 	}
 	// +kubebuilder:scaffold:builder
