+++ internal/controller/openclaw_instance.go
@@ -1,0 +1,5 @@
+package controller
+
+const (
+	OpenClawInstanceName = "OpenClawInstance"
+)
