+++ openspec/changes/archive/2026-04-02-split-controllers/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-04-02
